Move corpse_creation into combat package
