+++ combat/corpse_creation.py
