+++ utils/__init__.py
@@ -11,8 +11,8 @@ try:  # optional during early initialization
         auto_calc,
         auto_calc_secondary,
         make_corpse,
-        spawn_corpse,
     )
+    from combat.corpse_creation import spawn_corpse
 except Exception:  # pragma: no cover - may fail before Django setup
     pass
 try:
